src/python/s4_event_readers/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
colorist==1.1.2
    # via -r requirements.in
construct==2.10.70
    # via
    #   -r requirements.in
    #   etl-parser
defusedxml==0.7.1
    # via dissect-etl
dissect-cstruct==4.6
    # via dissect-etl
dissect-etl==3.13
    # via -r requirements.in
dissect-util==3.22
    # via dissect-etl
etl-parser==1.0.1
    # via -r requirements.in
pycryptodome==3.23.0
    # via -r requirements.in
wheel==0.45.1
    # via -r requirements.in
